requirements.txt
aiohttp==3.4.4
multidict==4.4.0
pytest==3.8.0
pytest-arraydiff==0.2
pytest-asyncio==0.9.0
pytest-cov==2.6.0
pytest-openfiles==0.3.0
pytest-remotedata==0.3.0
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
